25/08/2024 7pm

+++ database/migrations/2024_06_16_160025_create_user_data_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->string('phone');
             $table->integer('country_id');
             $table->integer('state_id')->nullable();
-            $table->unsignedBigInteger('city_id')->nullable();
+            $table->integer('city_id')->nullable();
             $table->string('city')->nullable();
             $table->string('address')->nullable();
